backend/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
django-filter==22.1
django-jsonfield==1.4.1
djangorestframework==3.14.0
filelock==3.9.0
platformdirs==3.1.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.20.0
djangorestframework-simplejwt
Pillow
django-cors-headers==3.7.0